Share a couple common test functions
