+++ tests/cli/test_api_gen.py
@@ -1,4 +1,3 @@
-import os
 from datetime import datetime
 from pathlib import Path
 from tempfile import TemporaryDirectory
@@ -7,20 +6,9 @@ from unittest import mock
 import pytest
 
 from openapi_spec_tools.cli.api_gen import generate_api
+from tests.cli.helpers import read_text
 from tests.helpers import StringIo
 from tests.helpers import asset_filename
-
-
-@pytest.fixture
-def temp_working_dir():
-    with TemporaryDirectory() as temp_dir:
-        os.chdir(temp_dir)
-        yield temp_dir
-
-
-def _read_text(filename: str) -> str:
-    with open(filename, "r", encoding="utf-8", newline="\n") as fp:
-        return fp.read()
 
 
 @pytest.mark.parametrize(
@@ -94,5 +82,5 @@ def test_api_generate_success_copyright(copyright_fixture):
     }
     for fname in filenames:
         file = code_dir / fname
-        text = _read_text(file.as_posix())
+        text = read_text(file.as_posix())
         assert copyright_text in text
